Plantbot (#1831)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds the Plantbot!
The Plantbot will help keep Hydroponics Trays filled with water and
cleared of weeds.
It will add water in 10u steps if there is less than 80u in the tray,
and will clear weeds with 10% per step if there are more than 10% weeds
in the tray.
If the bot is destroyed, it will spill a bunch of water.

It does **not** add nutrients, sample, harvest or plant.

If emagged, it'll drink water from trays until no water remains.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/5ad42d3f-d708-414d-b38d-e3ffb2a60cdf)

https://github.com/user-attachments/assets/9e626078-41da-423e-a9e1-f8712323ca69

</p>
</details>

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Added Plantbot

---------

Signed-off-by: Timfa <timfalken@hotmail.com>
Co-authored-by: sleepyyapril <123355664+sleepyyapril@users.noreply.github.com>